Initial commit
An expansion of the changeinventorytabs mod.  All errors should now be fixed, added tabbing between traders on the traders screen.  The ability to switch between the different tabs that each trader has.  The ability to open the inventory screen, trader screen, flea screen, chat and the hideout with hotkeys from the main menu.